+++ src/main/java/tornadofx/Builders.kt
@@ -14,6 +14,7 @@ import javafx.scene.layout.*
 import javafx.scene.text.Text
 import javafx.util.Callback
 import javafx.util.StringConverter
+import tornadofx.control.DatePickerTableCell
 import java.time.LocalDate
 import java.util.concurrent.Callable
 import kotlin.reflect.KFunction
@@ -258,6 +259,19 @@ fun <S, T> TableColumn<S, T?>.makeComboBox(items: ObservableList<T>, afterCommit
     cellFactory = ComboBoxTableCell.forTableColumn(items)
     setOnEditCommit {
         val property = it.tableColumn.getCellObservableValue(it.rowValue) as ObjectProperty<T?>
+        property.value = it.newValue
+        afterCommit?.invoke(it)
+    }
+    return this
+}
+
+/**
+ * Create an editable DatePicker TableCell. This control requires tornadofx-controls on the classpath.
+ */
+fun <S> TableColumn<S, LocalDate?>.makeDatePicker(converter: StringConverter<LocalDate>? = DatePickerTableCell.DefaultLocalDateConverter(), afterCommit: ((TableColumn.CellEditEvent<S, LocalDate?>) -> Unit)? = null): TableColumn<S, LocalDate?> {
+    cellFactory = DatePickerTableCell.forTableColumn(converter)
+    setOnEditCommit {
+        val property = it.tableColumn.getCellObservableValue(it.rowValue) as ObjectProperty<LocalDate?>
         property.value = it.newValue
         afterCommit?.invoke(it)
     }