+++ src/main/java/tornadofx/Component.kt
@@ -131,8 +131,6 @@ abstract class UIComponent : Component() {
         }
     }
 
-    val stage: Stage get() = root.scene.window as Stage
-    val hasStage = root.scene.window is Stage
 }
 
 abstract class Fragment : UIComponent() {